feat: Creating the base api calls for the rooms, movies and seats
- Creating the Movie and Room models and they're api views
which inslude the creation, update, deletion and get item or all items
- Creating the Seat model and api views. For now I'm just
making calls to get all seats in a specific room or a specific seat
in a room

The Movie and Room models need some changes. The goal is to able
to create the seat when creating a room, and assign a room to a movie
and not the other way.