feat: Adding the session and ticket views
Making changes on the session model to save session's time
in the time format. I've also added a post endpoint to create
the sessions.

Finally I created a new model for the tickets holding too
endpoints, one to create a ticket and a second one for getting
a ticket by id.

Currently some of the view aren't user friendly. the input data
is usually id values. So I would either make the call from the
client side with a json including the names with their id's
or get only the name from the client side and make a query
search in the database at the backend side of the app. The
first solution is more appealing since it can be integrated
with most of the views and it'll put less strain on the
backend side.